# openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation
  version: '1.0'
servers:
  - url: http://localhost:8080
paths:
...
  /ads/{id}/comments:
    get:
      tags:
        - Комментарии
      summary: 'Получение комментариев объявления'
      operationId: getComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comments'
        '401':
          description: Unauthorized
        '404':
          description: Not found

    post:
      tags:
        - Комментарии
      summary: 'Добавление комментария к объявлению'
      operationId: addComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateComment'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
  /ads/{adId}/comments/{commentId}:
    delete:
      tags:
        - Комментарии
      summary: 'Удаление комментария'
      operationId: deleteComment
      parameters:
        - name: adId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '404':
          description: Not found
    patch:
      tags:
        - Комментарии
      summary: 'Обновление комментария'
      operationId: updateComment
      parameters:
        - name: adId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateComment'
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Comment'
        '403':
          description: Forbidden
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
components:
  schemas:
...
    CreateOrUpdateComment:
      required:
        - text
      type: object
      properties:
        text:
          type: string
          description: 'текст комментария'
          minLength: 8
          maxLength: 64
...
    Comment:
      type: object
      properties:
        author:
          type: integer
          format: int32
          description: 'id автора комментария'
        authorImage:
          type: string
          description: 'ссылка на аватар автора комментария'
        authorFirstName:
          type: string
          description: 'имя создателя комментария'
        createdAt:
          type: integer
          format: int64
          description: 'дата и время создания комментария в миллисекундах с 00:00:00 01.01.1970'
        pk:
          type: integer
          format: int32
          description: 'id комментария'
        text:
          type: string
          description: 'текст комментария'
...
    Comments:
      type: object
      properties:
        count:
          type: integer
          format: int32
          description: 'общее количество комментариев'
        results:
          type: array
          items:
            $ref: '#/components/schemas/Comment'