# src/main/java/ru/skypro/homework/dto/CreateOrUpdateCommentDto.java
package ru.skypro.homework.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.Size;


@Data
@AllArgsConstructor
@NoArgsConstructor
public class CreateOrUpdateCommentDto {
    @Size(min = 8, max = 64)
    private String text;

}